Split Nerf client and server parts
