+++ Makefile
@@ -1,5 +1,6 @@
 GOMINVERSION = 1.16
 NERF_CMD_PATH = "./cmd/nerf"
+NERF_SERVER_CMD_PATH = "./cmd/nerf-server"
 GO111MODULE = on
 export GO111MODULE
 
@@ -34,11 +35,14 @@ proto:
 	go build github.com/golang/protobuf/protoc-gen-go
 	PATH="$(PWD):$(PATH)" protoc --go_out=plugins=grpc:. *.proto
 	rm protoc-gen-go
-bin:
+client:
 	@-go build -ldflags "$(LDFLAGS)" -o ./nerf ${NERF_CMD_PATH}
+server:
+	@-go build -ldflags "$(LDFLAGS)" -o ./nerf-server ${NERF_SERVER_CMD_PATH}
 clean:
 	rm -rf ./build
 	rm -f ./nerf
+	rm -f ./nerf-server
 
 .FORCE:
-.DEFAULT_GOAL := bin
+.DEFAULT_GOAL := client
