+++ cmd/nerf/main.go
@@ -3,8 +3,6 @@ package main
 import (
 	"context"
 	"flag"
-	"fmt"
-	"net"
 	"os"
 	"os/signal"
 	"path"
@@ -17,82 +15,11 @@ import (
 	"google.golang.org/grpc"
 )
 
-func stringToLogLevel(level string) zapcore.Level {
-	switch string(level) {
-	case "debug", "DEBUG":
-		return zapcore.DebugLevel
-	case "info", "INFO":
-		return zapcore.InfoLevel
-	case "warn", "WARN":
-		return zapcore.WarnLevel
-	case "error", "ERROR":
-		return zapcore.ErrorLevel
-	}
-
-	return zapcore.InfoLevel
-}
-
-func startServer(lightHouse string) {
-	if lightHouse == "" {
-		fmt.Println("-lighthouse flag must be set")
-		flag.Usage()
-		os.Exit(1)
-	}
-
-	lightHouseIPS := strings.Split(lightHouse, ":")
-	if len(lightHouseIPS) < 2 {
-		fmt.Println("The format for lighthouse must be <NebulaIP>:<PublicIP>")
-		flag.Usage()
-		os.Exit(1)
-	}
-
-	if err := net.ParseIP(lightHouseIPS[0]); err == nil {
-		fmt.Println("NebulaIP address is not IPv4")
-		flag.Usage()
-		os.Exit(1)
-	}
-
-	if err := net.ParseIP(lightHouseIPS[1]); err == nil {
-		fmt.Println("PublicIP address is not IPv4")
-		flag.Usage()
-		os.Exit(1)
-	}
-
-	nerf.Cfg.Nebula.LightHouse.NebulaIP = lightHouseIPS[0]
-	nerf.Cfg.Nebula.LightHouse.PublicIP = lightHouseIPS[1]
-
-	nerf.Cfg.Logger.Debug("Nerf server started", zap.String("lightHouse", lightHouse))
-
-	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", 9000))
-	if err != nil {
-		nerf.Cfg.Logger.Fatal("failed to listen gRPC server", zap.Error(err))
-	}
-
-	grpcServer := grpc.NewServer()
-	nerf.RegisterServerServer(grpcServer, &nerf.Server{})
-
-	go func() {
-		for range time.Tick(10 * time.Second) {
-			if (time.Now().Unix() - nerf.Cfg.Teams.UpdatedAt) > 3600 {
-				nerf.Cfg.Logger.Debug(
-					"begin-of-sync Github Teams with local cache")
-				nerf.SyncTeams()
-				nerf.Cfg.Logger.Debug(
-					"end-of-sync Github Teams with local cache")
-			}
-		}
-	}()
-
-	if err := grpcServer.Serve(lis); err != nil {
-		nerf.Cfg.Logger.Fatal("can't serve gRPC", zap.Error(err))
-	}
-}
-
 func startClient(redirect bool) {
 	err := nerf.NebulaDownload()
 	if err != nil {
 		if _, err := os.Stat(nerf.NebulaExecutable()); err != nil {
-			nerf.Cfg.Logger.Fatal("can't install Nebula:", zap.Error(err))
+			nerf.ClientCfg.Logger.Fatal("can't install Nebula:", zap.Error(err))
 		}
 	}
 
@@ -101,27 +28,27 @@ func startClient(redirect bool) {
 	nerf.Auth()
 	e := nerf.GetFastestEndpoint()
 	if e.RemoteHost == "" {
-		nerf.Cfg.Logger.Fatal("no available gRPC endpoints found")
+		nerf.ClientCfg.Logger.Fatal("no available gRPC endpoints found")
 	}
 
-	nerf.Cfg.Logger.Debug("found fastest endpoint",
+	nerf.ClientCfg.Logger.Debug("found fastest endpoint",
 		zap.String("RemoteIP", e.RemoteIP),
 		zap.String("RemoteHost", e.RemoteHost),
 		zap.String("Description", e.Description))
 
 	if err := nerf.NebulaAddLightHouseStaticRoute(&e); err != nil {
-		nerf.Cfg.Logger.Fatal(
+		nerf.ClientCfg.Logger.Fatal(
 			"can't create route",
 			zap.String("destination", e.RemoteIP),
 			zap.Error(err),
 		)
 	}
 
-	nerf.Cfg.Logger.Debug("authorized", zap.String("login", nerf.Cfg.Login))
+	nerf.ClientCfg.Logger.Debug("authorized", zap.String("login", nerf.ClientCfg.Login))
 
 	conn, err := grpc.Dial(e.RemoteHost+":9000", grpc.WithInsecure())
 	if err != nil {
-		nerf.Cfg.Logger.Fatal(
+		nerf.ClientCfg.Logger.Fatal(
 			"can't connect to gRPC server",
 			zap.Error(err),
 			zap.String("remoteHost", e.RemoteHost),
@@ -133,13 +60,13 @@ func startClient(redirect bool) {
 
 	client := nerf.NewServerClient(conn)
 
-	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
+	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
 	defer cancel()
 
-	request := &nerf.Request{Token: &nerf.Cfg.Token, Login: &nerf.Cfg.Login}
+	request := &nerf.Request{Token: &nerf.ClientCfg.Token, Login: &nerf.ClientCfg.Login}
 	response, err := client.Connect(ctx, request)
 	if err != nil {
-		nerf.Cfg.Logger.Fatal(
+		nerf.ClientCfg.Logger.Fatal(
 			"can't connect to gRPC server",
 			zap.Error(err),
 			zap.String("remoteHost", e.RemoteHost),
@@ -147,7 +74,7 @@ func startClient(redirect bool) {
 		)
 	}
 
-	nerf.Cfg.Logger.Debug("connected to LightHouse",
+	nerf.ClientCfg.Logger.Debug("connected to LightHouse",
 		zap.String("ClientIP", *response.ClientIP),
 		zap.String("LightHouseIP", *response.LightHouseIP),
 		zap.Strings("Teams", response.Teams))
@@ -182,11 +109,11 @@ func startClient(redirect bool) {
 
 	out, err := os.Create(path.Join(nerf.NebulaDir(), "config.yml"))
 	if err != nil {
-		nerf.Cfg.Logger.Fatal("can't create Nebula config", zap.Error(err))
+		nerf.ClientCfg.Logger.Fatal("can't create Nebula config", zap.Error(err))
 	}
 
 	if _, err := out.WriteString(config); err != nil {
-		nerf.Cfg.Logger.Fatal("can't write Nebula config", zap.Error(err))
+		nerf.ClientCfg.Logger.Fatal("can't write Nebula config", zap.Error(err))
 	}
 	defer out.Close()
 
@@ -194,30 +121,32 @@ func startClient(redirect bool) {
 	signal.Notify(done, os.Interrupt)
 
 	if err := nerf.NebulaSetNameServers(&e, []string{*response.LightHouseIP}, true); err != nil {
-		nerf.Cfg.Logger.Fatal("can't set custom DNS servers", zap.Error(err))
+		nerf.ClientCfg.Logger.Fatal("can't set custom DNS servers", zap.Error(err))
 	}
 
 	if err := nerf.NebulaStart(); err != nil {
-		nerf.Cfg.Logger.Fatal("can't start Nebula client", zap.Error(err))
+		nerf.ClientCfg.Logger.Fatal("can't start Nebula client", zap.Error(err))
 	}
 
 	<-done
-	notify, err := client.Disconnect(ctx, &nerf.Notify{Login: &nerf.Cfg.Login})
+
+	ctx, cancel = context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+
+	notify, err := client.Disconnect(ctx, &nerf.Notify{Login: &nerf.ClientCfg.Login})
 	if err != nil {
-		nerf.Cfg.Logger.Error(
+		nerf.ClientCfg.Logger.Error(
 			"disconnect",
-			zap.String("Login", nerf.Cfg.Login),
+			zap.String("Login", nerf.ClientCfg.Login),
 			zap.String("Response", notify.String()),
 		)
 	}
-	if err = nerf.NebulaSetNameServers(&e, nerf.Cfg.SavedNameServers, false); err != nil {
-		nerf.Cfg.Logger.Fatal("can't revert name servers", zap.Error(err))
+	if err = nerf.NebulaSetNameServers(&e, nerf.ClientCfg.SavedNameServers, false); err != nil {
+		nerf.ClientCfg.Logger.Fatal("can't revert name servers", zap.Error(err))
 	}
 }
 
 func main() {
-	server := flag.Bool("server", false, "Start gRPC server to generate config for Nebula")
-	lightHouse := flag.String("lighthouse", "", "Set the lighthouse. E.g.: <NebulaIP>:<PublicIP>")
 	logLevel := flag.String(
 		"log-level",
 		"info",
@@ -233,26 +162,22 @@ func main() {
 		os.Exit(0)
 	}
 
-	nerf.Cfg = nerf.NewConfig()
+	nerf.ClientCfg = nerf.NewClientConfig()
 
 	logger, _ := zap.Config{
 		Encoding:    "json",
-		Level:       zap.NewAtomicLevelAt(stringToLogLevel(*logLevel)),
+		Level:       zap.NewAtomicLevelAt(nerf.StringToLogLevel(*logLevel)),
 		OutputPaths: []string{"stdout"},
 		EncoderConfig: zapcore.EncoderConfig{
 			MessageKey: "message",
 		},
 	}.Build()
 
-	nerf.Cfg.Logger = logger
+	nerf.ClientCfg.Logger = logger
 
 	defer func() {
-		_ = nerf.Cfg.Logger.Sync()
+		_ = nerf.ClientCfg.Logger.Sync()
 	}()
 
-	if *server {
-		startServer(*lightHouse)
-	} else {
-		startClient(*redirectAll)
-	}
+	startClient(*redirectAll)
 }
